- Add esbuild as a dependency and update build script - moved classes to own file
