+++ index.html
@@ -32,6 +32,6 @@
             <!--Netzplan kommt hierhin-->
         </div>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-        <script src="dist/srcipt.js"></script>
+        <script src="./dist/script.js" defer></script>
     </body>
 </html>
